refactor: AI 테스트 스크립트 모듈 분리 및 raw/corrected 결과 분리
Made-with: Cursor
